requirements.txt
google-cloud-storage
supabase
pandas
numpy
pillow
python-dotenv